+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Add support for [dry-monads](https://github.com/dry-rb/dry-monads).
+
 ## [0.3.0] (2018-03-08)
 
 ### Added